+++ tests/testthat/test-corpus.R
@@ -487,3 +487,32 @@ test_that("[.corpus out of bounds generates expected error", {
     expect_error(corp1[2], "Subscript out of bounds")
 })
 
+test_that("corpus printing works with new textual summary", {
+    expect_silent(
+        print(data_corpus_irishbudget2010, ndoc = 0, nchar = 0, show.summary = FALSE)
+    )
+    expect_output(
+        print(data_corpus_irishbudget2010, ndoc = 0, nchar = 0, show.summary = TRUE),
+        "Corpus consisting of 14 documents and 6 docvars.", fixed = TRUE
+    )
+    expect_output(
+        print(data_corpus_irishbudget2010, ndoc = 2, nchar = 10, show.summary = TRUE),
+        "Corpus consisting of 14 documents and 6 docvars.
+  [Lenihan, Brian (FF)] When I pre ...
+ [Bruton, Richard (FG)] This draco ...
+and 12 more documents.", fixed = TRUE
+    )
+    expect_output(
+        print(data_corpus_irishbudget2010, ndoc = 2, nchar = 10, show.summary = FALSE),
+        "  [Lenihan, Brian (FF)] When I pre ...
+ [Bruton, Richard (FG)] This draco ...
+and 12 more documents.", fixed = TRUE
+    )
+    expect_output(
+        print(data_corpus_irishbudget2010[1:2], ndoc = 3, nchar = 10, show.summary = TRUE),
+        "Corpus consisting of 2 documents and 6 docvars.
+  [Lenihan, Brian (FF)] When I pre ...
+ [Bruton, Richard (FG)] This draco ...", fixed = TRUE
+    )
+    
+})